requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
PyGithub>=2.1.0
pdfplumber>=0.10.0
-e .
